backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pypdf
python-multipart
litellm
qdrant-client
python-dotenv
boto3
passlib[bcrypt]
python-jose[cryptography]
python-docx
bcrypt
argon2-cffi
azure-identity
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
google-cloud-compute